Refactor connectivity into own module, add diagonal and M connectivities
