+++ src/lib.rs
@@ -64,3 +64,4 @@ pub mod template_matching;
 pub mod union_find;
 #[cfg(feature = "display-window")]
 pub mod window;
+pub mod connectivity;
